+++ src/core/meshcop/meshcop.cpp
@@ -317,7 +317,7 @@ exit:
 
 #if OPENTHREAD_FTD
 Error GeneratePskc(const char *                  aPassPhrase,
-                   const Mac::NetworkName &      aNetworkName,
+                   const MeshCoP::NetworkName &  aNetworkName,
                    const MeshCoP::ExtendedPanId &aExtPanId,
                    Pskc &                        aPskc)
 {
