+++ src/core/thread/thread_netif.cpp
@@ -83,6 +83,7 @@ ThreadNetif::ThreadNetif(Instance &aInstance)
     , mActiveDataset(aInstance)
     , mPendingDataset(aInstance)
     , mExtendedPanIdManager(aInstance)
+    , mNetworkNameManager(aInstance)
     , mIp6Filter(aInstance)
     , mKeyManager(aInstance)
     , mLowpan(aInstance)
